added tests for eclipseCenter, separated code from data on eclipsecenter.
