+++ utils/rpc/eclipse-2017/checkStations.js
@@ -3,7 +3,7 @@ const fs = require('fs'),
     rp = require('request-promise'),
     Q = require('q'),
     _ = require('lodash'),
-    eclipsePath = require('./eclipsePathCenter'),
+    eclipsePath = require('./eclipsePath').center,
     stationUtils = require('../../../src/server/rpc/procedures/eclipse-2017/stations.js'),
     csv = require('fast-csv'),
     Storage = require('../../../src/server/storage/storage'),
